Protect against bad unicode when serializing error payload
Since we started putting the whole bad batch's payload into the error
topic as JSON, we became susceptible to crashing when trying to put bad
unicode in. This patch protects against this issue by replacing bad
codepoints with the replacement character so we can at least see
something bad was there.